Sketch: add basic 2D sketch of rover frame
This sketch will be used to determine the shape of the rover's
drivetrain.

It shall be a Rocker-Bogie (like NASA mars Rovers).

Since the sketch is barely finished, not much info has been
gleaned from it. The whole reason I'm doing a sketch is so
that I can find out these basic dims before I do real 3D work.